--- satellitePropagatorExamples/SatellitePropagatorExamples/singlePerturbedSatellitePropagatorJSON.cpp
@@ -1,221 +1,0 @@
-/*    Copyright (c) 2010-2017, Delft University of Technology
- *    All rigths reserved
- *
- *    This file is part of the Tudat. Redistribution and use in source and
- *    binary forms, with or without modification, are permitted exclusively
- *    under the terms of the Modified BSD license. You should have received
- *    a copy of the license with this file. If not, please or visit:
- *    http://tudat.tudelft.nl/LICENSE.
- */
-
-#include <Tudat/SimulationSetup/tudatSimulationHeader.h>
-
-#include "SatellitePropagatorExamples/applicationOutput.h"
-
-//! Execute propagation of orbit of Asterix around the Earth.
-int main( int argc, char* argv[] )
-{
-    if ( argc != 2 )
-    {
-        std::cerr << "Usage: spsp \"relative or absolute path to a JSON input file\"" << std::endl;
-        return EXIT_FAILURE;
-    }
-    else
-    {
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        ///////////////////////            USING STATEMENTS              //////////////////////////////////////////////////////
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-        using namespace tudat;
-        using namespace tudat::json_interface;
-        using namespace tudat::simulation_setup;
-        using namespace tudat::propagators;
-        using namespace tudat::numerical_integrators;
-        using namespace tudat::orbital_element_conversions;
-        using namespace tudat::basic_mathematics;
-        using namespace tudat::gravitation;
-        using namespace tudat::numerical_integrators;
-
-
-        // Get input file path
-        std::string inputFilePath = argv[1];
-
-        // Set up and run the simulation
-        Simulation simulation( inputFilePath );
-        simulation.run( );
-
-        // Export the results
-        simulation.exportResults( );
-
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        ///////////////////////     CREATE ENVIRONMENT AND VEHICLE       //////////////////////////////////////////////////////
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-
-        // Set simulation time settings.
-        const double simulationStartEpoch = settings.getSimulationStartEpoch();
-        const double simulationEndEpoch = settings.getSimulationEndEpoch();
-
-        // Define body settings for simulation.
-        std::vector< std::string > bodiesToCreate;
-        bodiesToCreate.push_back( "Sun" );
-        bodiesToCreate.push_back( "Earth" );
-        bodiesToCreate.push_back( "Moon" );
-        bodiesToCreate.push_back( "Mars" );
-        bodiesToCreate.push_back( "Venus" );
-
-        // Create body objects.
-        std::map< std::string, boost::shared_ptr< BodySettings > > bodySettings =
-                getDefaultBodySettings( bodiesToCreate, simulationStartEpoch - 300.0, simulationEndEpoch + 300.0 );
-        for( unsigned int i = 0; i < bodiesToCreate.size( ); i++ )
-        {
-            bodySettings[ bodiesToCreate.at( i ) ]->ephemerisSettings->resetFrameOrientation( "J2000" );
-            bodySettings[ bodiesToCreate.at( i ) ]->rotationModelSettings->resetOriginalFrame( "J2000" );
-        }
-        NamedBodyMap bodyMap = createBodies( bodySettings );
-
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        ///////////////////////             CREATE VEHICLE            /////////////////////////////////////////////////////////
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-        // Create spacecraft object.
-        bodyMap[ "Asterix" ] = boost::make_shared< simulation_setup::Body >( );
-        bodyMap[ "Asterix" ]->setConstantBodyMass( 400.0 );
-
-        // Create aerodynamic coefficient interface settings.
-        double referenceArea = 4.0;
-        double aerodynamicCoefficient = 1.2;
-        boost::shared_ptr< AerodynamicCoefficientSettings > aerodynamicCoefficientSettings =
-                boost::make_shared< ConstantAerodynamicCoefficientSettings >(
-                    referenceArea, aerodynamicCoefficient * Eigen::Vector3d::UnitX( ), 1, 1 );
-
-        // Create and set aerodynamic coefficients object
-        bodyMap[ "Asterix" ]->setAerodynamicCoefficientInterface(
-                    createAerodynamicCoefficientInterface( aerodynamicCoefficientSettings, "Asterix" ) );
-
-        // Create radiation pressure settings
-        double referenceAreaRadiation = 4.0;
-        double radiationPressureCoefficient = 1.2;
-        std::vector< std::string > occultingBodies;
-        occultingBodies.push_back( "Earth" );
-        boost::shared_ptr< RadiationPressureInterfaceSettings > asterixRadiationPressureSettings =
-                boost::make_shared< CannonBallRadiationPressureInterfaceSettings >(
-                    "Sun", referenceAreaRadiation, radiationPressureCoefficient, occultingBodies );
-
-        // Create and set radiation pressure settings
-        bodyMap[ "Asterix" ]->setRadiationPressureInterface(
-                    "Sun", createRadiationPressureInterface(
-                        asterixRadiationPressureSettings, "Asterix", bodyMap ) );
-
-
-        // Finalize body creation.
-        setGlobalFrameBodyEphemerides( bodyMap, "SSB", "J2000" );
-
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        ///////////////////////            CREATE ACCELERATIONS          //////////////////////////////////////////////////////
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-        // Define propagator settings variables.
-        SelectedAccelerationMap accelerationMap;
-        std::vector< std::string > bodiesToPropagate;
-        std::vector< std::string > centralBodies;
-
-        // Define propagation settings.
-        std::map< std::string, std::vector< boost::shared_ptr< AccelerationSettings > > > accelerationsOfAsterix;
-        accelerationsOfAsterix[ "Earth" ].push_back( boost::make_shared< SphericalHarmonicAccelerationSettings >( 5, 5 ) );
-
-        accelerationsOfAsterix[ "Sun" ].push_back( boost::make_shared< AccelerationSettings >(
-                                                       basic_astrodynamics::central_gravity ) );
-        accelerationsOfAsterix[ "Moon" ].push_back( boost::make_shared< AccelerationSettings >(
-                                                        basic_astrodynamics::central_gravity ) );
-        accelerationsOfAsterix[ "Mars" ].push_back( boost::make_shared< AccelerationSettings >(
-                                                        basic_astrodynamics::central_gravity ) );
-        accelerationsOfAsterix[ "Venus" ].push_back( boost::make_shared< AccelerationSettings >(
-                                                         basic_astrodynamics::central_gravity ) );
-        accelerationsOfAsterix[ "Sun" ].push_back( boost::make_shared< AccelerationSettings >(
-                                                       basic_astrodynamics::cannon_ball_radiation_pressure ) );
-        accelerationsOfAsterix[ "Earth" ].push_back( boost::make_shared< AccelerationSettings >(
-                                                         basic_astrodynamics::aerodynamic ) );
-
-        accelerationMap[  "Asterix" ] = accelerationsOfAsterix;
-        bodiesToPropagate.push_back( "Asterix" );
-        centralBodies.push_back( "Earth" );
-
-        basic_astrodynamics::AccelerationMap accelerationModelMap = createAccelerationModelsMap(
-                    bodyMap, accelerationMap, bodiesToPropagate, centralBodies );
-
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        ///////////////////////             CREATE PROPAGATION SETTINGS            ////////////////////////////////////////////
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-        // Set Keplerian elements for Asterix.
-        Eigen::Vector6d asterixInitialStateInKeplerianElements;
-        asterixInitialStateInKeplerianElements( semiMajorAxisIndex ) = 7500.0E3;
-        asterixInitialStateInKeplerianElements( eccentricityIndex ) = 0.1;
-        asterixInitialStateInKeplerianElements( inclinationIndex ) = unit_conversions::convertDegreesToRadians( 85.3 );
-        asterixInitialStateInKeplerianElements( argumentOfPeriapsisIndex )
-                = unit_conversions::convertDegreesToRadians( 235.7 );
-        asterixInitialStateInKeplerianElements( longitudeOfAscendingNodeIndex )
-                = unit_conversions::convertDegreesToRadians( 23.4 );
-        asterixInitialStateInKeplerianElements( trueAnomalyIndex ) = unit_conversions::convertDegreesToRadians( 139.87 );
-
-        double earthGravitationalParameter = bodyMap.at( "Earth" )->getGravityFieldModel( )->getGravitationalParameter( );
-        const Eigen::Vector6d asterixInitialState = convertKeplerianToCartesianElements(
-                    asterixInitialStateInKeplerianElements, earthGravitationalParameter );
-
-
-        boost::shared_ptr< TranslationalStatePropagatorSettings< double > > propagatorSettings =
-                boost::make_shared< TranslationalStatePropagatorSettings< double > >
-                ( centralBodies, accelerationModelMap, bodiesToPropagate, asterixInitialState, simulationEndEpoch );
-
-        const double fixedStepSize = 10.0;
-        boost::shared_ptr< IntegratorSettings< > > integratorSettings =
-                boost::make_shared< IntegratorSettings< > >
-                ( rungeKutta4, 0.0, fixedStepSize );
-
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        ///////////////////////             PROPAGATE ORBIT            ////////////////////////////////////////////////////////
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-
-        // Create simulation object and propagate dynamics.
-        SingleArcDynamicsSimulator< > dynamicsSimulator(
-                    bodyMap, integratorSettings, propagatorSettings, true, false, false );
-        std::map< double, Eigen::VectorXd > integrationResult = dynamicsSimulator.getEquationsOfMotionNumericalSolution( );
-
-
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        ///////////////////////        PROVIDE OUTPUT TO CONSOLE AND FILES           //////////////////////////////////////////
-        ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-
-
-        Eigen::VectorXd finalIntegratedState = (--integrationResult.end( ) )->second;
-        // Print the position (in km) and the velocity (in km/s) at t = 0.
-        std::cout << "Single Earth-Orbiting Satellite Example." << std::endl <<
-                     "The initial position vector of Asterix is [km]:" << std::endl <<
-                     asterixInitialState.segment( 0, 3 ) / 1E3 << std::endl <<
-                     "The initial velocity vector of Asterix is [km/s]:" << std::endl <<
-                     asterixInitialState.segment( 3, 3 ) / 1E3 << std::endl;
-
-        // Print the position (in km) and the velocity (in km/s) at t = 86400.
-        std::cout << "After " << simulationEndEpoch <<
-                     " seconds, the position vector of Asterix is [km]:" << std::endl <<
-                     finalIntegratedState.segment( 0, 3 ) / 1E3 << std::endl <<
-                     "And the velocity vector of Asterix is [km/s]:" << std::endl <<
-                     finalIntegratedState.segment( 3, 3 ) / 1E3 << std::endl;
-
-        // Write perturbed satellite propagation history to file.
-        input_output::writeDataMapToTextFile( integrationResult,
-                                              "singlePerturbedSatellitePropagationHistory.dat",
-                                              tudat_applications::getOutputPath( ),
-                                              "",
-                                              std::numeric_limits< double >::digits10,
-                                              std::numeric_limits< double >::digits10,
-                                              "," );
-        // Final statement.
-        // The exit code EXIT_SUCCESS indicates that the program was successfully executed.
-        return EXIT_SUCCESS;
-    }
-
-}
